+++ CHANGELOG.md
@@ -48,6 +48,9 @@ Commands annotations have also been rewritten to support AOT compilation and be 
 - Fix command aliases not being associated with their names in Bukkit's command system
 - Fix commands being added to `Root` even when unable to be registered in Bukkit's command system
 - Remove `Commands.remove(CommandNode<T>, String...)` - this method was almost never used and posed a technical burden, use `Commands.remove(CommandNode<T>, String)` instead
+- Remove `Commands.resolve(Object)` - this method have been made redundant by the rewritten command annotations
+- Remove `Commands.resolve(Object, String)` - this method have been made redundant by the rewritten command annotations
+- Remove `Nodes.Builder.then(Object, String)` - this method have been made redundant by the rewritten command annotations
 
 **Command Annotations**
 Commands annotations have been rewritten to support AOT compilation and be less verbose.Please refer to the wiki for